refactor: move taxonomy classification to separate module and update usage
